+++ src/test/java/kr/hhplus/be/server/domain/coupon/service/CouponServiceIntegrationTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.domain.coupon.service;
 
-import kr.hhplus.be.server.IntegrationTestSupport;
+import kr.hhplus.be.server.support.IntegrationTestSupport;
 import kr.hhplus.be.server.domain.coupon.dto.info.IssuedCouponInfo;
 import kr.hhplus.be.server.domain.coupon.entity.Coupon;
 import kr.hhplus.be.server.domain.coupon.entity.IssuedCoupon;
