+++ src/test/java/kr/hhplus/be/server/domain/product/service/ProductStockConcurrenyTest.java
@@ -4,7 +4,7 @@ import kr.hhplus.be.server.domain.product.dto.StockCommand;
 import kr.hhplus.be.server.domain.product.entity.ProductStock;
 import kr.hhplus.be.server.domain.product.repository.ProductStockRepository;
 import kr.hhplus.be.server.domain.support.exception.CustomException;
-import kr.hhplus.be.server.IntegrationTestSupport;
+import kr.hhplus.be.server.support.IntegrationTestSupport;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
